+++ CHANGELOG.md
@@ -19,5 +19,6 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Launching a server on Windows from a folder with a space not working [#544]
 - Account UUID without dashes not being censored in logs [#546]
 - Add some more protection to exceptions when logging out debug network logs
+- Change Update Data button to Check For Updates to make it clearer what it does
 
 ### Misc